cleaned up the TitleManager class and removed the original InventoryManager class
