+++ src/io/teamelite/brush_menu/utilities/TitleManager.java
@@ -1,0 +1,41 @@
+package io.teamelite.brush_menu.utilities;
+
+import net.minecraft.server.v1_8_R2.IChatBaseComponent;
+import net.minecraft.server.v1_8_R2.PacketPlayOutTitle;
+import org.bukkit.ChatColor;
+import org.bukkit.craftbukkit.v1_8_R2.entity.CraftPlayer;
+import org.bukkit.entity.Player;
+
+/**
+ * @name 		BrushMenu
+ * @author 		Liam Reffell and Kieron Wiltshire
+ * @contact 	http://www.mcteamelite.com/
+ * @license 	MIT License
+ * @description
+ * 				The plugin requires Java 1.6 or higher and depends on the VoxelSniper
+ * 				Bukkit plugin. It allows a user to open up an inventory interface and
+ * 				select their desired VoxelSniper brush.
+ */
+public class TitleManager {
+
+    public static void createTitle(Player p, String t, int fIT, int sT, int fOT, ChatColor c) {
+        IChatBaseComponent chatTitle = IChatBaseComponent.ChatSerializer.a("{\"text\": \"" + t + "\",color:" + c.name().toLowerCase() + "}");
+
+        PacketPlayOutTitle title = new PacketPlayOutTitle(PacketPlayOutTitle.EnumTitleAction.TITLE, chatTitle);
+        PacketPlayOutTitle length = new PacketPlayOutTitle(fIT, sT, fOT);
+
+        ((CraftPlayer) p).getHandle().playerConnection.sendPacket(title);
+        ((CraftPlayer) p).getHandle().playerConnection.sendPacket(length);
+    }
+
+    public static void createSubTitle(Player p, String t, int fIT, int sT, int fOT, ChatColor c) {
+        IChatBaseComponent chatTitle = IChatBaseComponent.ChatSerializer.a("{\"text\": \"" + t + "\",color:" + c.name().toLowerCase() + "}");
+
+        PacketPlayOutTitle title = new PacketPlayOutTitle(PacketPlayOutTitle.EnumTitleAction.SUBTITLE, chatTitle);
+        PacketPlayOutTitle length = new PacketPlayOutTitle(fIT, sT, fOT);
+
+        ((CraftPlayer) p).getHandle().playerConnection.sendPacket(title);
+        ((CraftPlayer) p).getHandle().playerConnection.sendPacket(length);
+    }
+
+}
